refactor(UP): impl-split SFUpgradeExecutionService across 2 .cpp [BUILD-PASSED]
Slice UP, as a pure impl-split (not the card's careful batch-repair extraction). Batch connection
repair (Save/Fix connection refs across destroy/respawn + manifest capture/validate) moves to
SFUpgradeExecutionService_ConnectionRepair.cpp (649). The 3 file-scope static connector-finders ride
along (their only users are in that block). SFUpgradeExecutionService.cpp 2,537 -> 1,939. Zero
behavior change. Shipping PackagePlugin clean (ExitCode=0). Solo lane.
